Restore previous Clang tests (with version 12)
Make the newer Clang a separate test (test-14 for qemu and test-28
for qemuv8) and restore the previous tests with Clang 12 (test-11
and test-21 respectively). This allows testing with both the older
and newer versions. Note that the qemu test-14 fails at "xtest
pkcs11_1000": the TA prints "E/TA:  stack smashing detected" and
panics.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>